requirements.txt
asgiref==3.2.9
Django==3.0.7
django-cors-headers==3.4.0
django-restframework==0.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
